Move `package-info.java` to java package.
